+++ packages/pass/components/Item/Containers/ItemNew.tsx
@@ -1,7 +1,8 @@
-import type { FC } from 'react';
+import { type FC } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory, useParams } from 'react-router-dom';
 
+import { usePassCore } from '@proton/pass/components/Core/PassCoreProvider';
 import { AliasNew } from '@proton/pass/components/Item/Alias/Alias.new';
 import { CreditCardNew } from '@proton/pass/components/Item/CreditCard/CreditCard.new';
 import { LoginNew } from '@proton/pass/components/Item/Login/Login.new';
@@ -12,8 +13,6 @@ import type { ItemNewViewProps } from '@proton/pass/components/Views/types';
 import { itemCreationIntent } from '@proton/pass/store/actions';
 import { selectDefaultVault, selectVaultLimits } from '@proton/pass/store/selectors';
 import type { ItemCreateIntent, ItemType } from '@proton/pass/types';
-
-import { usePassCore } from '../../Core/PassCoreProvider';
 
 const itemNewMap: { [T in ItemType]: FC<ItemNewViewProps<T>> } = {
     login: LoginNew,