+++ packages/pass/store/sagas/invites/invite-accept.saga.ts
@@ -18,18 +18,24 @@ import { selectInviteByToken } from '@proton/pass/store/selectors/invites';
 import type { Invite, ItemRevision, Maybe, Share, ShareGetResponse, ShareType } from '@proton/pass/types';
 import noop from '@proton/utils/noop';
 
+type AcceptInviteChannel = RequestProgress<ItemRevision[], null>;
+
 function* acceptInviteWorker({ payload, meta: { request } }: ReturnType<typeof inviteAcceptIntent>) {
+    const requestId = request.id;
+    const { inviteToken } = payload;
+
     try {
         yield put(stopEventPolling());
-        const invite: Maybe<Invite> = yield select(selectInviteByToken(payload.inviteToken));
+
+        const invite: Maybe<Invite> = yield select(selectInviteByToken(inviteToken));
         if (!invite) throw new Error(c('Error').t`Unknown invite`);
 
         const encryptedShare: ShareGetResponse = yield acceptInvite({ ...payload, inviteKeys: invite.keys });
         const share: Maybe<Share<ShareType.Vault>> = yield parseShareResponse(encryptedShare);
         if (!share) throw new Error(c('Error').t`Could not open invited vault`);
 
-        const progressChannel = eventChannel<RequestProgress<ItemRevision[]>>((emitter) => {
-            requestItemsForShareId(share.shareId, (value) => emitter({ type: 'progress', value }))
+        const progressChannel = eventChannel<AcceptInviteChannel>((emitter) => {
+            requestItemsForShareId(share.shareId, (progress) => emitter({ type: 'progress', progress, data: null }))
                 .then((result) => emitter({ type: 'done', result }))
                 .catch((error) => emitter({ type: 'error', error }))
                 .finally(() => emitter(END));
@@ -38,21 +44,13 @@ function* acceptInviteWorker({ payload, meta: { request } }: ReturnType<typeof i
         });
 
         while (true) {
-            const action: RequestProgress<ItemRevision[]> = yield take(progressChannel);
-            switch (action.type) {
-                case 'progress':
-                    if (invite.vault.itemCount === 0) break;
-                    yield put(requestProgress(request.id, action.value));
-                    break;
-                case 'done':
-                    yield put(inviteAcceptSuccess(request.id, payload.inviteToken, share, action.result));
-                    break;
-                case 'error':
-                    throw action.error;
-            }
+            const action: AcceptInviteChannel = yield take(progressChannel);
+            if (action.type === 'progress') yield put(requestProgress(requestId, action.progress));
+            if (action.type === 'done') yield put(inviteAcceptSuccess(requestId, inviteToken, share, action.result));
+            if (action.type === 'error') throw action.error;
         }
     } catch (err) {
-        yield put(inviteAcceptFailure(request.id, err));
+        yield put(inviteAcceptFailure(requestId, err));
     } finally {
         yield put(startEventPolling());
     }