+++ packages/pass/store/sagas/vaults/vault-move-all-items.saga.ts
@@ -1,27 +1,39 @@
-import { put, select, takeEvery } from 'redux-saga/effects';
+import { put, select, take, takeEvery } from 'redux-saga/effects';
 
-import { moveItems } from '@proton/pass/lib/items/item.requests';
 import {
     vaultMoveAllItemsFailure,
     vaultMoveAllItemsIntent,
+    vaultMoveAllItemsProgress,
     vaultMoveAllItemsSuccess,
 } from '@proton/pass/store/actions';
+import { type BulkMoveItemsChannel, bulkMoveChannel } from '@proton/pass/store/sagas/items/item-bulk-move.saga';
 import { selectItemsByShareId } from '@proton/pass/store/selectors';
 import type { RootSagaOptions } from '@proton/pass/store/types';
 import type { ItemRevision } from '@proton/pass/types';
 
 function* moveAllItemsWorker(
     { onItemsUpdated }: RootSagaOptions,
-    { payload: { shareId, content, destinationShareId }, meta }: ReturnType<typeof vaultMoveAllItemsIntent>
+    { payload, meta }: ReturnType<typeof vaultMoveAllItemsIntent>
 ) {
-    try {
-        const items: ItemRevision[] = yield select(selectItemsByShareId(shareId));
-        const movedItems = (yield moveItems(items, destinationShareId)) as ItemRevision[];
+    const { shareId, content, destinationShareId } = payload;
+    const itemsToMove: ItemRevision[] = yield select(selectItemsByShareId(shareId));
+    const channel = bulkMoveChannel(itemsToMove, destinationShareId);
 
-        yield put(vaultMoveAllItemsSuccess(meta.request.id, { shareId, destinationShareId, content, movedItems }));
-        onItemsUpdated?.();
-    } catch (e) {
-        yield put(vaultMoveAllItemsFailure(meta.request.id, { shareId, content }, e));
+    while (true) {
+        const action: BulkMoveItemsChannel = yield take(channel);
+
+        if (action.type === 'progress') {
+            yield put(
+                vaultMoveAllItemsProgress(meta.request.id, action.progress, {
+                    ...action.data,
+                    destinationShareId,
+                })
+            );
+            onItemsUpdated?.();
+        }
+
+        if (action.type === 'done') yield put(vaultMoveAllItemsSuccess(meta.request.id, { content }));
+        if (action.type === 'error') yield put(vaultMoveAllItemsFailure(meta.request.id, payload, action.error));
     }
 }