+++ packages/pass/types/data/items.dto.ts
@@ -1,6 +1,6 @@
 import type { CustomAliasCreateRequest } from '../api';
 import type { ItemType } from '../protobuf';
-import type { Item } from './items';
+import type { IndexedByShareIdAndItemId, Item } from './items';
 
 type AliasMailbox = {
     id: number;
@@ -56,3 +56,7 @@ export type ItemImportIntent<T extends ItemType = ItemType> = Item<T, ItemImport
     createTime?: number;
     modifyTime?: number;
 };
+
+/** This data-structure does not uses lists to avoid
+ * iterations when checking for inclusions  */
+export type BulkSelectionDTO = IndexedByShareIdAndItemId<true>;